+++ lib/highway/service.ml
@@ -32,6 +32,7 @@ let make_simple ?middleware ~route handler =
 ;;
 
 let dispatch
+      ?(decode = false)
       ~given_method
       ~given_path
       ~given_query_params
@@ -49,7 +50,9 @@ let dispatch
       :: others ->
       if Route.has_method given_method route && precondition request
       then (
-        match Route.extract_values ~given_path ~given_query_params route with
+        match
+          Route.extract_values ~decode ~given_path ~given_query_params route
+        with
         | Some (args, param) ->
           if postcondition args param request
           then (