+++ CHANGES.md
@@ -1,3 +1,9 @@
+### Unreleased
+
+#### Highway 
+
+- Add `pct` encoding and decoding for path segment and query components
+
 ### v0.0.1 (Paris)
 
 #### Highway